backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.0
dnspython==2.4.2
google-generativeai==0.3.2
Pillow>=9.0.0
openai>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv==1.0.0